Drop support for Python2.7 and Python3.5
