+++ .travis.yml
@@ -5,10 +5,6 @@ cache:
   - pip
 matrix:
   include:
-    - python: 2.7
-      env: TOXENV=py27
-    - python: 3.5
-      env: TOXENV=py35
     - python: 3.6
       env: TOXENV=py36
     - python: 3.7
@@ -16,11 +12,10 @@ matrix:
     - python: 3.8
       env: TOXENV=py38
 install:
-  - pip install tox codecov awscli
+  - pip install tox awscli
 before_script:
   - ./scripts/test_data/upload_test_data.sh
 script:
   - tox
-after_success:
-  # - ./scripts/test_data/delete_test_data.sh
-  - codecov
+# after_success:
+#   - ./scripts/test_data/delete_test_data.sh
