+++ Pipfile
@@ -14,8 +14,8 @@ awscli = "*"
 tox = "*"
 twine = "*"
 wheel = "*"
-pandas = ">=0.24.0"
-pyarrow = ">=0.15.0"
+pandas = ">=1.0.0"
+pyarrow = ">=1.0.0"
 SQLAlchemy = ">=1.0.0, <2.0.0"
 pytest = ">=3.5"
 pytest-cov = "*"
