+++ pyathena/async_cursor.py
@@ -1,10 +1,8 @@
 # -*- coding: utf-8 -*-
-from __future__ import absolute_import, unicode_literals
-
 import logging
 from concurrent.futures.thread import ThreadPoolExecutor
+from multiprocessing import cpu_count
 
-from pyathena import cpu_count
 from pyathena.common import CursorIterator
 from pyathena.cursor import BaseCursor
 from pyathena.error import NotSupportedError, ProgrammingError
