+++ pyathena/cursor.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
-from __future__ import absolute_import, unicode_literals
-
 import logging
 
 from pyathena.common import BaseCursor, CursorIterator
