+++ pyathena/pandas/async_cursor.py
@@ -1,12 +1,10 @@
 # -*- coding: utf-8 -*-
-from __future__ import absolute_import, unicode_literals
-
 import logging
+from multiprocessing import cpu_count
 
-from pyathena import cpu_count
 from pyathena.async_cursor import AsyncCursor
 from pyathena.common import CursorIterator
-from pyathena.result_set import AthenaPandasResultSet
+from pyathena.pandas.result_set import AthenaPandasResultSet
 
 _logger = logging.getLogger(__name__)
