+++ pyathena/sqlalchemy_athena.py
@@ -1,12 +1,9 @@
 # -*- coding: utf-8 -*-
-from __future__ import absolute_import, unicode_literals
-
 import math
 import numbers
 import re
 
 import tenacity
-from future.utils import raise_from
 from sqlalchemy import exc, util
 from sqlalchemy.engine import Engine, reflection
 from sqlalchemy.engine.default import DefaultDialect
@@ -422,7 +419,7 @@ class AthenaDialect(DefaultDialect):
             ]
         except OperationalError as e:
             if not self._retry_if_data_catalog_exception(e, schema, table_name):
-                raise_from(NoSuchTableError(table_name), e)
+                raise NoSuchTableError(table_name) from e
             else:
                 raise e
