+++ tests/pandas/test_async_pandas_cursor.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
-from __future__ import absolute_import, unicode_literals
-
 import contextlib
 import random
 import string
@@ -9,18 +7,16 @@ import unittest
 from datetime import datetime
 from random import randint
 
-from past.builtins.misc import xrange
-
-from pyathena.async_cursor import AsyncCursor
 from pyathena.error import NotSupportedError, ProgrammingError
 from pyathena.model import AthenaQueryExecution
+from pyathena.pandas.async_cursor import AsyncPandasCursor
 from pyathena.result_set import AthenaResultSet
 from tests import WithConnect
 from tests.conftest import ENV, S3_PREFIX, SCHEMA
 from tests.util import with_async_pandas_cursor
 
 
-class TestAsyncCursor(unittest.TestCase, WithConnect):
+class TestAsyncPandasCursor(unittest.TestCase, WithConnect):
     @with_async_pandas_cursor()
     def test_fetchone(self, cursor):
         query_id, future = cursor.execute("SELECT * FROM one_row")
@@ -61,7 +57,7 @@ class TestAsyncCursor(unittest.TestCase, WithConnect):
         self.assertEqual(result_set.fetchall(), [(1,)])
         query_id, future = cursor.execute("SELECT a FROM many_rows ORDER BY a")
         result_set = future.result()
-        self.assertEqual(result_set.fetchall(), [(i,) for i in xrange(10000)])
+        self.assertEqual(result_set.fetchall(), [(i,) for i in range(10000)])
 
     @with_async_pandas_cursor()
     def test_iterator(self, cursor):
@@ -207,7 +203,7 @@ class TestAsyncCursor(unittest.TestCase, WithConnect):
         self.assertEqual(df.shape[0], 10000)
         self.assertEqual(df.shape[1], 1)
         self.assertEqual(
-            [(row["a"],) for _, row in df.iterrows()], [(i,) for i in xrange(10000)]
+            [(row["a"],) for _, row in df.iterrows()], [(i,) for i in range(10000)]
         )
 
     @with_async_pandas_cursor()
@@ -231,12 +227,12 @@ class TestAsyncCursor(unittest.TestCase, WithConnect):
 
     def test_open_close(self):
         with contextlib.closing(self.connect()) as conn:
-            with conn.cursor(AsyncCursor):
+            with conn.cursor(AsyncPandasCursor):
                 pass
 
     def test_no_ops(self):
         conn = self.connect()
-        cursor = conn.cursor(AsyncCursor)
+        cursor = conn.cursor(AsyncPandasCursor)
         self.assertRaises(
             NotSupportedError, lambda: cursor.executemany("SELECT * FROM one_row", [])
         )
@@ -246,7 +242,7 @@ class TestAsyncCursor(unittest.TestCase, WithConnect):
     @with_async_pandas_cursor()
     def test_empty_result(self, cursor):
         table = "test_pandas_cursor_empty_result_" + "".join(
-            [random.choice(string.ascii_lowercase + string.digits) for _ in xrange(10)]
+            [random.choice(string.ascii_lowercase + string.digits) for _ in range(10)]
         )
         location = "{0}{1}/{2}/".format(ENV.s3_staging_dir, S3_PREFIX, table)
         query_id, future = cursor.execute(
