+++ tests/test_formatter.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
-from __future__ import absolute_import, unicode_literals
-
 import textwrap
 import unittest
 from datetime import date, datetime
