+++ tox.ini
@@ -1,26 +1,16 @@
 [tox]
-envlist = py27,py35,py36,py37,py38
+skipsdist = True
+envlist = py36,py37,py38
 
 [testenv]
 deps =
-    SQLAlchemy>=1.0.0, <2.0.0
-    pandas>=0.24.0
-    pyarrow>=0.15.0
-    pytest>=3.5
-    pytest-cov
-    pytest-flake8>=1.0.1
-    pytest-black
-    pytest-isort
-    pytest-xdist
+    pipenv
+    codecov
+setenv =
+    PYTHONPATH={toxinidir}
+    PIPENV_VERBOSITY=-1
 commands =
-    pytest --cov pyathena --cov-report html --cov-report term --flake8 --black --isort
-passenv =
-    AWS_*
-
-[testenv:py{27,35}]
-deps =
-    {[testenv]deps}
-commands =
-    pytest --cov pyathena --cov-report html --cov-report term --flake8 --black
-passenv =
-    {[testenv]passenv}
+    pipenv sync --dev
+    pipenv run pytest --cov pyathena --cov-report html --cov-report term --flake8 --black --isort
+    codecov -e TOXENV
+passenv = AWS_* TOXENV CI TRAVIS TRAVIS_* CODECOV_*
